Fixed several splicing and path-fitting bugs and made further filters.

+++ src/StartNXSimulation.py
@@ -47,8 +47,9 @@ from copy import copy
 import opaque.maps.gen_icp as gen_icp
 import atexit
 def cleanup():
-	for p in gen_icp.overlapPool:
-		p.terminate()
+	if gen_icp.overlapPool != None:
+		for p in gen_icp.overlapPool:
+			p.terminate()
 atexit.register(cleanup)
